🚚 Move style files
- styles 폴더 만들어 전역 스타일 파일 모두 추가
- 추후 vanilla extract로 모두 변경 예정
