+++ src/App.tsx
@@ -1,6 +1,7 @@
-import './App.css';
 import { buttonThemeClass } from './Component/Button/theme.css';
 import Router from './Router';
+import './styles/App.css';
+import './styles/svg.css';
 
 function App() {
   return (
