+++ src/styles/App.css
@@ -1,0 +1,5 @@
+.App {
+  width: 100vw;
+  height: 100vh;
+  text-align: center;
+}
