+++ src/styles/reset.css
@@ -1,0 +1,76 @@
+html,
+body,
+h1,
+input,
+select {
+  font-family: 'Apple SD Gothic Neo', arial, sans-serif;
+}
+body,
+h1 {
+  font-size: 14px;
+}
+h1 {
+  font-weight: normal;
+  margin: 0;
+  padding: 0;
+}
+h3 {
+  font-weight: normal;
+  margin: 0;
+  padding: 0;
+  font-size: 20px;
+  line-height: 1.3;
+}
+body {
+  margin: 0;
+  background: #fff;
+  color: #202124;
+}
+a {
+  color: inherit;
+  text-decoration: none;
+}
+button {
+  border: none;
+  background: inherit;
+  box-shadow: none;
+  border-radius: 0;
+  margin: 0;
+  padding: 0;
+  overflow: visible;
+  cursor: pointer;
+}
+ol li {
+  list-style: none;
+}
+ol,
+ul,
+li {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+input {
+  border: none;
+  font-size: 14px;
+}
+em {
+  font-weight: bold;
+  font-style: normal;
+}
+table {
+  border-collapse: collapse;
+  border-spacing: 0;
+}
+
+textarea {
+  border: none;
+  overflow: auto;
+  outline: none;
+
+  -webkit-box-shadow: none;
+  -moz-box-shadow: none;
+  box-shadow: none;
+
+  resize: none; /*remove the resize handle on the bottom right*/
+}
