+++ src/styles/svg.css
@@ -1,0 +1,4 @@
+svg path {
+  fill: white;
+  stroke: white;
+}
